# Minor rewrite of hashing function to improve performance by ~3x