--- a/package.json
+++ b/package.json
@@ -62,6 +62,7 @@
   "scripts": {
     "test": "npm run test-ts && npm run test-unit -- --ci --coverage && npm run build && npm run test-perf",
     "test-perf": "NODE_ENV=production node benchmarks/perf.js",
+    "test-perf-hash": "NODE_ENV=production node benchmarks/perf-hash.js",
     "test-unit": "jest --setupFiles ./tests/setup.js --roots ./src packages && npm run test-addon-global && npm run test-addon-prefixer",
     "test-addon-global": "cd global && npm run test",
     "test-addon-prefixer": "cd prefixer && npm run test",