--- a/src/core/to-hash.js
+++ b/src/core/to-hash.js
@@ -7,5 +7,8 @@
  * @param {String} str
  * @returns {String}
  */
-export let toHash = (str) =>
-    'go' + str.split('').reduce((out, i) => (101 * out + i.charCodeAt(0)) >>> 0, 11);
+export let toHash = (str) => {
+    let i = 0, l = str.length, out = 11;
+    while (i < l) out = (101 * out + str.charCodeAt(i++)) >>> 0
+    return 'go' + out;
+}