Goober's current hashing function causes unnecessary memory allocations (`Array.split()`) as well as compute cycles / function calls (`Array.reduce()`). The following simple change does not increase the bundle size, and yields much better performance (at least 2x faster, nearly 3x):

```diff
-export let toHash = (str) =>
-    'go' + str.split('').reduce((out, i) => (101 * out + i.charCodeAt(0)) >>> 0, 11);
+export let toHash = (str) => {
+    let i = 0, l = str.length, out = 11;
+    while (i < l) out = (101 * out + str.charCodeAt(i++)) >>> 0
+    return 'go' + out;
+}
```

Comparisons using Benchmark.js:

`npm run test-perf-hash`

=>

```
> NODE_ENV=production node benchmarks/perf-hash.js


Starting HASH!
▸ goober original HASH x 3,180,521 ops/sec ±1.06% (87 runs sampled)
▸ goober optimized HASH x 8,219,240 ops/sec ±1.36% (84 runs sampled)

Fastest is: goober optimized HASH
```

```
> NODE_ENV=production node benchmarks/perf-hash.js


Starting HASH!
▸ goober original HASH x 3,582,669 ops/sec ±0.92% (90 runs sampled)
▸ goober optimized HASH x 9,067,875 ops/sec ±1.42% (94 runs sampled)

Fastest is: goober optimized HASH
```


```
> NODE_ENV=production node benchmarks/perf-hash.js

Starting HASH!
▸ goober original HASH x 3,517,773 ops/sec ±1.02% (89 runs sampled)
▸ goober optimized HASH x 10,189,264 ops/sec ±1.81% (90 runs sampled)

Fastest is: goober optimized HASH
```

```
> NODE_ENV=production node benchmarks/perf-hash.js

Starting HASH!
▸ goober original HASH x 3,368,369 ops/sec ±1.45% (85 runs sampled)
▸ goober optimized HASH x 9,699,551 ops/sec ±1.49% (90 runs sampled)

Fastest is: goober optimized HASH
```