@danielweck 2021-03-24
Side note: you'll be pleased to know that Goober's hashing algorithm (pre-optimization) is already the fastest of the bunch ... naturally the downside is an increased risk of collision compared with other hashing schemes (and I suspect not as good distribution too). Do you have a reference / source for the technique used in Goober?

PS: based on my (very) basic tests, Twind's "cyrb32" implementation seems to hit the right spot:

https://gist.github.com/danielweck/fb48cafa2e6543e81f313e961ea8496c

CC @sastan