@danielweck 2021-03-24
Update: I added a `Benchmark.js` suite in this PR:

`npm run test-perf-hash`

=>

```
> NODE_ENV=production node benchmarks/perf-hash.js


Starting HASH!
▸ goober original HASH x 3,180,521 ops/sec ±1.06% (87 runs sampled)
▸ goober optimized HASH x 8,219,240 ops/sec ±1.36% (84 runs sampled)

Fastest is: goober optimized HASH
```