@cristianbote 2021-03-24
@danielweck this is freaking awesome! Love it! Thanks for the benchmark as well.

`goober`'s hash-ing algo was based on Java's string hash. It was/is the best size/perf ratio I could figure out at that moment. Of course, nothing is forever and everything is up for change. I remember talking with @jviide about it as well and he pointed out the same collision argument, IIRC.

Now, there's one other bench that I would like to run before settling on this, and that is the browser one 😃  cause these are running on node runtime. So, I went ahead and copied the above test in https://esbench.com/bench/605b608d6c89f600a57004d3 and oh my. You are right. This is by far the fastest.

The only other bench that doesn't get _that_ much perf boost is this one:
![Screenshot 2021-03-24 at 18 10 46](https://user-images.githubusercontent.com/3405161/112343936-46902180-8ccc-11eb-960e-7a3a8b35ca5e.png)

The size increase it is around `+16B` which I believe is worth it.