@sastan 2021-03-24
It seems the twind version is faster. I do not know about the size impact.

![Bildschirmfoto 2021-03-24 um 18 37 54](https://user-images.githubusercontent.com/514405/112357783-219dad80-8cd0-11eb-8dc3-b3f5c35c4fe0.png)


https://esbench.com/bench/605b608d6c89f600a57004d3