@danielweck 2021-03-24
Ah, just a word of caution, in your ESBench I noticed an error that comes from my own benchmark (but not in Goober's actual code, otherwise the unit tests wouldn't pass!)
See the change of place for `++` increment in this commit:

https://github.com/cristianbote/goober/pull/271/commits/479d367f53cd77ef01b07b0a5dd16b85014eca89

I re-run `npm run test-perf-hash` just to be sure :)

```
> NODE_ENV=production node benchmarks/perf-hash.js

Starting HASH!
▸ goober original HASH x 3,517,773 ops/sec ±1.02% (89 runs sampled)
▸ goober optimized HASH x 10,189,264 ops/sec ±1.81% (90 runs sampled)

Fastest is: goober optimized HASH
```

```
> NODE_ENV=production node benchmarks/perf-hash.js

Starting HASH!
▸ goober original HASH x 3,368,369 ops/sec ±1.45% (85 runs sampled)
▸ goober optimized HASH x 9,699,551 ops/sec ±1.49% (90 runs sampled)

Fastest is: goober optimized HASH
```