@danielweck 2021-03-24
> It seems the twind version is faster.

Strange ... I added Twind's hash routine to the benchmark, it is much slower:

# Goober original
```js
    const c = 'go' + str.split('').reduce((out, s) => (101 * out + s.charCodeAt(0)) >>> 0, 11);
```

# Goober optimized
```js
    let i = 0, l = str.length, out = 11;
    while (i < l) out = (101 * out + str.charCodeAt(i++)) >>> 0
    const c = 'go' + out;
```

# Twind (`cyrb32` with `Math.imul()`)
```js
    let i = str.length - 1, out = 9;
    while (i >= 0) out = Math.imul(out ^ str.charCodeAt(i--), 0x5f356495);
    const c = 'go' + ((out ^ (out >>> 9)) >>> 0).toString(36);
```

# RESULTS

`npm run test-perf-hash`
=>
```
> NODE_ENV=production node benchmarks/perf-hash.js

Starting HASH!
▸ goober original HASH x 3,512,636 ops/sec ±1.03% (91 runs sampled)
▸ goober optimized HASH x 10,347,402 ops/sec ±1.02% (90 runs sampled)
▸ twind HASH x 2,458,171 ops/sec ±0.70% (96 runs sampled)

Fastest is: goober optimized HASH
```
```
> NODE_ENV=production node benchmarks/perf-hash.js


Starting HASH!
▸ goober original HASH x 3,464,524 ops/sec ±1.61% (85 runs sampled)
▸ goober optimized HASH x 10,256,783 ops/sec ±1.07% (92 runs sampled)
▸ twind HASH x 2,453,381 ops/sec ±0.80% (94 runs sampled)

Fastest is: goober optimized HASH
```