@danielweck 2021-03-24
Wow, ESBench running in the browser (as opposed to Benchmark.js running in Node) => HUUUGE difference: Twind wins in this case! (as @sastan said) I guess `Math.imul()` is not optimized in NodeJS? (note that I am using Firefox)

<img width="443" alt="Screenshot 2021-03-24 at 19 04 58" src="https://user-images.githubusercontent.com/625201/112369278-eac99680-8cd3-11eb-9d38-536ceef768ab.png">


```
const STR = `abcdefghijklmnopqrstuvwyxz ABCDEFGHIJKLMNOPQRSTUVWYXZ 0123456789`;

const original = (str) => {
    const c = 'go' + str.split('').reduce((out, s) => (101 * out + s.charCodeAt(0)) >>> 0, 11);
};

const optimized = (str) => {
    let i = 0, l = str.length, out = 11;
    while (i < l) out = (101 * out + str.charCodeAt(i++)) >>> 0
    const c = 'go' + out;
};

const twind = (str) => {
    let i = str.length - 1, out = 9;
    while (i >= 0) out = Math.imul(out ^ str.charCodeAt(i--), 0x5f356495);
    const c = 'go' + ((out ^ (out >>> 9)) >>> 0).toString(36);
};
```