@sastan 2021-03-24
That is interesting! I run that in Safari. Re-benching in Chrome confirm your results.

![Bildschirmfoto 2021-03-24 um 20 09 54](https://user-images.githubusercontent.com/514405/112369799-e786d880-8cdc-11eb-983e-17af08991215.png)

What we can say for sure is that the optimized version is faster than the current goober one.