@danielweck 2021-03-24
Safari same as Firefox (Twind / `Math.imul()` / `cyrb32` much faster)

<img width="563" alt="Screenshot 2021-03-24 at 19 09 46" src="https://user-images.githubusercontent.com/625201/112369989-c4582b00-8cd4-11eb-95c7-6ae378f1170c.png">