@danielweck 2021-03-24
...BUT, in Chrome Twind's algo is slow!

Go figure!?

<img width="495" alt="Screenshot 2021-03-24 at 19 12 09" src="https://user-images.githubusercontent.com/625201/112370153-f2d60600-8cd4-11eb-883b-f4de5635b525.png">

<img width="495" alt="Screenshot 2021-03-24 at 19 12 38" src="https://user-images.githubusercontent.com/625201/112370157-f49fc980-8cd4-11eb-9b1d-a1c854d7aeed.png">